requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
langdetect>=1.0.9
python-dateutil>=2.8.2
validators>=0.22.0
